refactor: extract astra-evolution crate (~1,802 LOC)
Extract 4 evolution modules from runtime to new astra-evolution crate:
- types.rs (363 LOC) — evolution signals, proposals, and axis types
- store.rs (661 LOC) — proposal persistence with file-based storage
- evolver.rs (356 LOC) — fast-path proposal generation
- signal_collector.rs (424 LOC) — signal detection from runtime events

All 51 evolution tests now run in astra-evolution crate.
Runtime re-exports via 1-line stubs for backward compatibility.
