+++ rust/Cargo.toml
@@ -19,6 +19,7 @@ members = [
     "crates/astra-messaging",
     "crates/astra-text-utils",
     "crates/astra-config",
+    "crates/astra-evolution",
     "crates/astra-learning",
 ]
 resolver = "2"
@@ -30,6 +31,7 @@ edition = "2024"
 [workspace.dependencies]
 astra-core = { path = "crates/core" }
 astra-config = { path = "crates/astra-config" }
+astra-evolution = { path = "crates/astra-evolution" }
 astra-learning = { path = "crates/astra-learning" }
 astra-messaging = { path = "crates/astra-messaging" }
 astra-text-utils = { path = "crates/astra-text-utils" }
